RVT Fix, Grass Adjustments and Rock Placement
In this branch, I ventured out to fix the issues with the RVT having compilation errors. I also adjusted the grass material to not cull the backfaces of the mesh, giving the grass more volume. I also created a Landscape Grass Type with the Rock Models to place them procedurally throughout the scene, at a lower scale. These rocks also benefit from the RVT, blending with the ground colour and normals.