New Rock Texture and Grass Shader Clean Up
Adjusted the Cliff Rocks texture to match the other rock assets by changing the stone colour from the saturated purple to grey. Also re-added the wave lines to the grass by passing it through to the emission colour at a low strength level. Did remove a lot of the old grass code but a lot of it went unused due to the RVT colour sampling. Can provide the old grass shader if needed!