MainMenu Buttons Glow Hover
Main menu buttons now increase in glow strength when hovered. Imported and added the studio logo to the main menu. Also imported the sfx for the UI button click/hover. This hasn't been implemented at this point, wanted to discuss whether it would be better as a hover sound or click sound.