Main Menu Button Text and Credit/Option Scenes Created
Added text to the buttons, set to no-hit so it doesn't disrupt button functionality. Created widget blueprints for the credits and options scenes and added button functionality to the main menu. Very makeshift scenes at the moment with a back button to return to main menu. Layout and art options in consideration currently for these scenes. Overall, unless font needs to be changed, the main menu is only awaiting the game logo/title.