Continued Forest Environment Development
Continued work on the forest environment. Some RVT colour map changes. Added the river material instance and updated the stylised water material with a normal rotation parameter to allow for a more dynamic material. Attempted experimentation for the cliffs to sample the RVT instead of using a ZUP texture sample, didn't work but further research will be conducted.