Added Tree Assets
Added branch and large tree 3D meshes and texture, although haven't made the materials yet. Had an issue with water of the river appearing over the character model in-game but resolved by changing the sort order of the water mesh to -1. Continued fleshing out the environment and added the large tree to the scene that will act as the entrance into Lucia's domain.