Foliage Development
Added the foliage master material and create the requirement material instances from that. The large tree assets now has foliage and the wisteria hanging from the tree. Flowers have been fixed from previous issue and are ready to paint on landscape. However before they are painted consider the light issue. The foliage mesh is appearing dark in particular locations, not showing the full texture only black. This may be a lighting issue or a particular setting. Explore this further before painting flowers as mesh changes may need to be implemented to combat this.