Quick UI Test
Imported and roughly implemented the character info pannel for testing. THIS IS NOT THE FINAL CHARACTER INFO PANNEL ASSET, it is just a test asset and will be refined further. At the moment I have implemented it with the Quest UI Widget for quick testing purposes, but that if subject to change.