Major Lighting Update
Experimented with alternate ways to light the scene. Added a Lightmasss Volume which is the most important part about this method. Ultimately, I was aiming to get a similar look to Lumens without using Lumens. This experimentation came about when I realised that the game was heavily taxing the GPU, and the currently changes have helped it. There is still room for further improvement. I will not continue with this lighting system now until the final build is being complete as it requires full asset overhauls and rebuilds.