Another UI Test
Like other the previous test, I quickly implemented the test assets for the HUD.  This time was testing whether the darker outlines looked better (which they did) so this should be the final/one of the last tests. Again the assets were placed on the Quest UI Blueprint for easier implementation, but that's subject to change.