Created Essence Collectable
I made a niagara system for the essence collectable. This required me to import specialised texture and create a unique material that communicates with niagara. I also created two niagara emitters for this niagara system. This collectable has momentarily be place in the level as a niagara system, but when it is turned into a collectable it'll go inside a blueprint and be visible is the level as a blueprint rather than niagara system.